+++ src/RayTracing/Light.java
@@ -4,13 +4,13 @@ package RayTracing;
 
 public class Light
 {
-    private int[] Position;
-    private int[] RGB;
+    private float[] Position;
+    private float[] RGB;
     private float specularIntensity;
     private float shadowIntensity;
     private float lightRadius;
 
-    public Light(int[] position, int[] RGB, float specularIntensity, float shadowIntensity, float lightRadius) {
+    public Light(float[] position, float[] RGB, float specularIntensity, float shadowIntensity, float lightRadius) {
         Position = position;
         this.RGB = RGB;
         this.specularIntensity = specularIntensity;
