+++ src/RayTracing/Material.java
@@ -3,7 +3,7 @@ package RayTracing;
 
 public class Material
 {
-    public Material(int[] diffuseColor, int[] specularColor, int[] reflectionColor, float phongSpecularityCoefficient, float transparency) {
+    public Material(float[] diffuseColor, float[] specularColor, float[] reflectionColor, float phongSpecularityCoefficient, float transparency) {
         DiffuseColor = diffuseColor;
         SpecularColor = specularColor;
         ReflectionColor = reflectionColor;
@@ -12,9 +12,9 @@ public class Material
     }
 
     //Properties
-    public int [] DiffuseColor;
-    public int [] SpecularColor;
-    public int [] ReflectionColor;
+    public float [] DiffuseColor;
+    public float [] SpecularColor;
+    public float [] ReflectionColor;
     public float PhongSpecularityCoefficient;
     public float Transparency;
