+++ src/RayTracing/Triangle.java
@@ -3,28 +3,28 @@ package RayTracing;
 
 public class Triangle
 {
-    private int[] vertex1;
-    private int[] vertex2;
-    private int[] vertex3;
+    private float[] vertex1;
+    private float[] vertex2;
+    private float[] vertex3;
     private int materialIndex;
 
 
-    public Triangle(int[] vertex1, int[] vertex2, int[] vertex3, int materialIndex) {
+    public Triangle(float[] vertex1, float[] vertex2, float[] vertex3, int materialIndex) {
         this.vertex1 = vertex1;
         this.vertex2 = vertex2;
         this.vertex3 = vertex3;
         this.materialIndex = materialIndex;
     }
 
-    public int[] getVertex1() {
+    public float[] getVertex1() {
         return vertex1;
     }
 
-    public int[] getVertex2() {
+    public float[] getVertex2() {
         return vertex2;
     }
 
-    public int[] getVertex3() {
+    public float[] getVertex3() {
         return vertex3;
     }
